refactor(agent): remove surface-specific leftovers
